Added Passthrough feature and Fixed bugs
