--- Assets/Scenes/prefabLab.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 74fe4dda561caa644a356b93a684f189
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
