--- Assets/Scripts/ClockDisplay.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b99045b76ac274ed8ba39dd7dd10c589
