--- Assets/Scripts/GameManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b5008f5d72aac0f4aae429d9310a715b
