+++ Assets/Scripts/GameManager.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e69202fd832f042ec8cbfc12ad2afcc4
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
