+++ Assets/Scripts/GameManager/GameManager.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b5008f5d72aac0f4aae429d9310a715b
