+++ Assets/Scripts/GameManager/NewObjectSpawner.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e6dbdd0efbe964a6184534bdebb8a959
