+++ Assets/Scripts/GameManager/SimpleObjectSpawner.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 072b38f88742c7742882bc477790c764
