--- Assets/Scripts/MagicScaler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 341c65059af6f48ce9b8dd107ece74c7
