--- Assets/Scripts/NewObjectSpawner.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e6dbdd0efbe964a6184534bdebb8a959
