+++ Assets/Scripts/PrefabManager/MagicScaler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 341c65059af6f48ce9b8dd107ece74c7
