+++ Assets/Scripts/PrefabManager/TwoHandPinchScaler.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: cf47559cfa195414598f85185055616a
