--- Assets/Scripts/SimpleObjectSpawner.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 072b38f88742c7742882bc477790c764
