--- Assets/Scripts/SpawnToggleHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 6c2bc3ba372e04f7eb3760afa0d049c6
