--- Assets/Scripts/TwoHandPinchScaler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: cf47559cfa195414598f85185055616a
