+++ Assets/Scripts/Utilities.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cec94a10384a64800a0892339670bc9c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
