+++ Assets/Textures.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9a0a98257e9a546858227a8ad76427e8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
