+++ Assets/Textures/Wall.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9d3fba41a8bc74606b0345c506d33491
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
